--- src/main/java/com/project/sooktoring/common/exception/EmptyFileException.java
@@ -1,7 +1,0 @@
-package com.project.sooktoring.common.exception;
-
-public class EmptyFileException extends RuntimeException {
-    public EmptyFileException(String message) {
-        super(message);
-    }
-}
