+++ src/main/java/com/project/sooktoring/exception/domain/user/auth/InvalidGoogleIdTokenException.java
@@ -1,4 +1,4 @@
-package com.project.sooktoring.user.auth.exception;
+package com.project.sooktoring.exception.domain.user.auth;
 
 public class InvalidGoogleIdTokenException extends RuntimeException {
