+++ src/main/java/com/project/sooktoring/exception/handler/ErrorCode.java
@@ -1,0 +1,13 @@
+package com.project.sooktoring.exception.handler;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorCode {
+
+    private final HttpStatus httpStatus;
+    private final String message;
+}
