+++ src/main/java/com/project/sooktoring/exception/handler/dto/ErrorResponse.java
@@ -1,0 +1,4 @@
+package com.project.sooktoring.exception.handler.dto;
+
+public class ErrorResponse {
+}
