+++ src/main/java/com/project/sooktoring/exception/handler/dto/MtrExResponse.java
@@ -1,5 +1,6 @@
-package com.project.sooktoring.common.exception.dto;
+package com.project.sooktoring.exception.handler.dto;
 
+import com.project.sooktoring.mentoring.dto.request.MtrRequest;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,11 +10,11 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class EnumExResponse {
+public class MtrExResponse {
 
     private int status;
 
     private String message;
 
-    private String source;
+    private MtrRequest request;
 }
