+++ src/main/java/com/project/sooktoring/mentoring/service/MentoringService.java
@@ -8,8 +8,8 @@ import com.project.sooktoring.mentoring.dto.response.MtrToResponse;
 import com.project.sooktoring.user.profile.domain.UserProfile;
 import com.project.sooktoring.mentoring.dto.request.MtrRequest;
 import com.project.sooktoring.mentoring.dto.request.MtrUpdateRequest;
-import com.project.sooktoring.mentoring.exception.MtrDuplicateException;
-import com.project.sooktoring.mentoring.exception.MtrTargetException;
+import com.project.sooktoring.exception.domain.mentoring.MtrDuplicateException;
+import com.project.sooktoring.exception.domain.mentoring.MtrTargetException;
 import com.project.sooktoring.mentoring.repository.MentoringRepository;
 import com.project.sooktoring.user.profile.repository.UserProfileRepository;
 import lombok.RequiredArgsConstructor;
