+++ src/main/java/com/project/sooktoring/user/auth/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.project.sooktoring.user.auth.controller;
 
 import com.project.sooktoring.user.auth.dto.request.AuthRequest;
 import com.project.sooktoring.user.auth.dto.request.TokenRequest;
-import com.project.sooktoring.common.exception.dto.AuthExResponse;
+import com.project.sooktoring.exception.handler.dto.AuthExResponse;
 import com.project.sooktoring.user.auth.dto.response.AuthResponse;
 import com.project.sooktoring.user.auth.dto.response.TokenResponse;
 import com.project.sooktoring.user.auth.service.AuthService;
